Resolución del ejercicio 21 de la parte 4

+++ part4/blogs/controllers/blogs.js
@@ -41,6 +41,26 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+
+  if(!token)
+  {
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+
+  const blog = await Blog.findById(request.params.id)
+  if(!blog)
+  {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if(blog.user.toString() !== decodedToken.id)
+  {
+    return response.status(403).json({ error: 'unauthorized action' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
